Remove all remaining references to "abstract spark"
